Moves pwv and transmission functions into pwv_trans module
