+++ pwv_kpno/__init__.py
@@ -32,18 +32,19 @@ An Incomplete Guide to Getting Started:
 
     To check what years data is locally available for:
 
-      >>> pwv_kpno.available_data()
+      >>> from pwv_kpno import pwv_trans
+      >>> pwv_trans.available_data()
 
 
     To update the locally available data with any new measurements:
 
-      >>> pwv_kpno.update_models()
+      >>> pwv_trans.update_models()
 
 
     To retrieve the atmospheric model for a line of sight PWV concentration of
     13.5 mm:
 
-      >>> pwv_kpno.transmission_pwv(13.5)
+      >>> pwv_trans.transmission_pwv(13.5)
 
 
     To retrieve the atmospheric model for a datetime:
@@ -58,34 +59,27 @@ An Incomplete Guide to Getting Started:
       >>>                      minute=35,
       >>>                      tzinfo=pytz.utc)
       >>>
-      >>> pwv_kpno.transmission(date=obsv_date, airmass=1.2)
+      >>> pwv_trans.transmission(date=obsv_date, airmass=1.2)
 
 
     To access the PWV measurements as an astropy table:
 
       >>> # All locally available PWV measurements
-      >>> pwv_kpno.measured_pwv()
+      >>> pwv_trans.measured_pwv()
       >>>
       >>> # All PWV measurements taken on November 14, 2016
-      >>> pwv_kpno.measured_pwv(year=2016, month=11, day=14)
+      >>> pwv_trans.measured_pwv(year=2016, month=11, day=14)
 
 
     To access the modeled PWV level at Kitt Peak as an astropy table:
 
       >>> # The entire model from 2010 to present
-      >>> pwv_kpno.modeled_pwv()
+      >>> pwv_trans.modeled_pwv()
       >>>
       >>> # The modeled PWV level only for November 14, 2016
-      >>> pwv_kpno.modeled_pwv(year=2016, month=11, day=14)
+      >>> pwv_trans.modeled_pwv(year=2016, month=11, day=14)
 """
 
-
-from .pwv_data import available_data
-from .pwv_data import measured_pwv
-from .pwv_data import modeled_pwv
-from .pwv_data import update_models
-from .transmission import transmission
-from .transmission import transmission_pwv
 
 __author__ = 'Daniel Perrefort'
 __copyright__ = 'Copyright 2017, Daniel Perrefort'
